assign4

+++ N01685558_Cumulative1/Cumulative1/Models/Teacher.cs
@@ -16,6 +16,6 @@
         public DateTime hiredate { get; set; }
         //Teacher salary
 
-        public double salary {  get; set; }    
+        public decimal salary {  get; set; }    
     }
 }
